Move comparator to its own namespace, register in bootstrap, use in a few places in tests.
