+++ tests/MoneyTest.php
@@ -48,6 +48,12 @@ final class MoneyTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(0 <= $result, $money->greaterThanOrEqual($other));
         $this->assertEquals(-1 === $result, $money->lessThan($other));
         $this->assertEquals(0 >= $result, $money->lessThanOrEqual($other));
+
+        if ($result === 0) {
+            $this->assertEquals($money, $other);
+        } else {
+            $this->assertNotEquals($money, $other);
+        }
     }
 
     /**
@@ -130,7 +136,7 @@ final class MoneyTest extends \PHPUnit_Framework_TestCase
         foreach ($allocated as $key => $money) {
             $compareTo = new Money($results[$key], $money->getCurrency());
 
-            $this->assertTrue($money->equals($compareTo));
+            $this->assertEquals($money, $compareTo);
         }
     }
 
@@ -147,7 +153,7 @@ final class MoneyTest extends \PHPUnit_Framework_TestCase
         foreach ($allocated as $key => $money) {
             $compareTo = new Money($results[$key], $money->getCurrency());
 
-            $this->assertTrue($money->equals($compareTo));
+            $this->assertEquals($money, $compareTo);
         }
     }
