+++ tests/bootstrap.php
@@ -1,0 +1,4 @@
+<?php
+
+\SebastianBergmann\Comparator\Factory::getInstance()
+    ->register(new \Money\PHPUnit\Comparator());
